Add a detailed comment about peformance in edge cases.
This script will handle certain common yaml parsing errors.
E.g., this script can properly handle a `- ---` caused by Logseq placing a bullet in front of the hyphens.
This is technically not a proper yaml header but is a common error so this script accounts for it.
However, certain rare edge cases, e.g. a `-` followed by a line break then a `- ---`,
or an open header marker with no close header,
will not be parsed. Details in comments.