Rewrite to enable automatic copying of files. Code works.
Currently does not have a --dry-run argument so there is not currently a way to get a report only as in the initial code.
Fully tested for 7 test cases including --verbose and --overwrite. Passed all tests.
The case of both --verbose and --overwrite together has not yet been tested.
Error output needs to be improved for Test Case 4 (YAML Parsing error)